f1 가중치 API

+++ src/main/java/org/example/backend/f1/team/F1TeamResponse.java
@@ -8,7 +8,7 @@ public record F1TeamResponse(
         String logoUrl,
         String videoUrl,
         Integer seasonPosition,
-        Integer seasonPoint,
+        Double seasonPoint,
         Integer seasonWins,
         Integer seasonPodiums,
         Integer careerWins,
